PlzDedicatedWorker: send COEP violation reports to an environment settings object owner
This CL makes a dedicated worker sends COEP violation reports to its
an environment settings object owner during the worker initialization
process when cross-origin-embedder-policy is invalid.
For PlzDedicatedWorker, the owner is the creator frame which starts
the worker directly.
For non-PlzDedicatedWorker, the owner is the nearest ancestor frame.
This is not align with the spec but keeps the current behavior.

Bug: 1060837
Change-Id: I088470e7355c94f48b6ac8d51b5cf813b9d181ad